+++ src/utils/mocks.js
@@ -23,6 +23,8 @@ import Various2 from "./demos/various2.vue?raw";
 import Various3 from "./demos/various3.vue?raw";
 import Various4 from "./demos/various4.vue?raw";
 import Various5 from "./demos/various5.vue?raw";
+import IconOnly from "./demos/iconOnly.vue?raw";
+import Loading from "./demos/loading.vue?raw";
 
 const clickHandler = () => {
   console.log("clicked");
@@ -101,6 +103,17 @@ export const demos = [
     },
     description: "Default button with custom style when is outlined.",
     html: CustomStyleOutlined,
+  },
+  {
+    id: 29,
+    title: "Loading",
+    propsData: {
+      buttonText: "Default",
+      isLoading: true,
+      onClick: clickHandler
+    },
+    description: "Default button with loading state.",
+    html: Loading,
   },
   {
     id: 9,
@@ -243,6 +256,18 @@ export const demos = [
     },
     description: "Default button with type icon outlined.",
     html: TypeIcon,
+  },
+  {
+    id: 27,
+    title: "Type Icon and icon only (star)",
+    propsData: {
+      type: "icon",
+      icon: "star",
+      iconOnly: true,
+      onClick: clickHandler,
+    },
+    description: "Default button with type icon and icon only.",
+    html: IconOnly,
   },
   {
     id: 19,